+++ R/mapping_result_class.R
@@ -10,13 +10,14 @@
 #' between the winning bin and the bulk.
 #' @param top_2_distance Decimal. The absolute difference between the best and
 #' second best mapping buckets. Higher indicates a less doubtful mapping.
-#' @param confident_mapping Boolean. TRUE when the mapped bin's lower
+#' @param strong_mapping Boolean. TRUE when the mapped bin's lower
 #' bound is higher than the maximum upper bound of the other bins.
 #' @param history A dataframe of the correlation score (decimal) and
 #' confidence bounds (decimal pairs) for each bin.
 #' Access with `mapping_history()`
 #' @param bootstrap_iterations Integer. The number of iterations used during
 #' the bootstrap.
+#' @param metric Character. The metric used to evaluate mappings.
 #'
 #' @return A MappingResult object
 #'
@@ -69,26 +70,32 @@
 #' best_bin(result)
 #' best_correlation(result)
 #' top_2_distance(result)
-#' confident_mapping(result)
+#' strong_mapping(result)
 #' mapping_history(result)
 #' bootstrap_iterations(result)
+#' metric(result)
+#'
+#' # Setters
+#' bulk_name(result) <- "New Name"
 MappingResult <- function(
     bulk_name,
     best_bin,
     best_correlation,
     top_2_distance,
-    confident_mapping,
+    strong_mapping,
     history,
-    bootstrap_iterations) {
+    bootstrap_iterations,
+    metric="spearman") {
 
   return(methods::new("MappingResult",
           bulk_name = bulk_name,
           best_bin = best_bin,
           best_correlation = best_correlation,
           top_2_distance = top_2_distance,
-          confident_mapping = confident_mapping,
+          strong_mapping = strong_mapping,
           history = history,
-          bootstrap_iterations = bootstrap_iterations
+          bootstrap_iterations = bootstrap_iterations,
+          metric = metric
   ))
 
 }
@@ -102,9 +109,10 @@ setClass(
         best_bin = "numeric",
         best_correlation = "numeric",
         top_2_distance = "numeric",
-        confident_mapping = "logical",
+        strong_mapping = "logical",
         history = "data.frame",
-        bootstrap_iterations = "numeric"
+        bootstrap_iterations = "numeric",
+        metric = "character"
     )
 )
 
@@ -136,8 +144,8 @@ setMethod(
                 " top_2_distance=", top_2_distance(object)
             ),
             paste(
-                "\t Confident Result:",
-                confident_mapping(object),
+                "\t Strong Result:",
+                strong_mapping(object),
                 "(next max upper ",
                 non_top_mapping_best_upper_bound,
                 ")"
@@ -174,6 +182,32 @@ setMethod(
     f = "bulk_name",
     signature = "MappingResult",
     definition = function(x) x@bulk_name
+)
+
+#' @title Set name of bulk of a BLASE Mapping Results object.
+#'
+#' @concept mapping-result-object
+#'
+#' @rdname bulk_name-setter
+#' @param x a [MappingResult] object
+#' @param value String. The name of the bulk used to map against.
+#'
+#' @returns Nothing
+#'
+#' @importFrom methods validObject
+#' @export
+#' @inherit BlaseData-class examples
+setGeneric("bulk_name<-", function(x, value) standardGeneric("bulk_name<-"))
+
+#' @rdname bulk_name-setter
+setMethod(
+  f = "bulk_name<-",
+  signature = "MappingResult",
+  definition = function(x, value) {
+    x@bulk_name <- value
+    validObject(x)
+    x
+  }
 )
 
 
@@ -236,25 +270,25 @@ setMethod(
     definition = function(x) x@top_2_distance
 )
 
-#' @title Get if the result is confident for a BLASE Mapping Results object.
+#' @title Get if the result is strong for a BLASE Mapping Results object.
 #'
 #' @concept mapping-result-object
 #'
-#' @rdname mapping-result-confident-mapping-getter
+#' @rdname mapping-result-strong-mapping-getter
 #' @param x a [MappingResult] object
-#' @returns Boolean. TRUE if the result is confident, otherwise FALSE
+#' @returns Boolean. TRUE if the result is strong, otherwise FALSE
 #' @export
 #' @inherit MappingResult examples
 setGeneric(
-    "confident_mapping",
-    function(x) standardGeneric("confident_mapping")
+    "strong_mapping",
+    function(x) standardGeneric("strong_mapping")
 )
 
-#' @rdname mapping-result-confident-mapping-getter
+#' @rdname mapping-result-strong-mapping-getter
 setMethod(
-    f = "confident_mapping",
+    f = "strong_mapping",
     signature = "MappingResult",
-    definition = function(x) x@confident_mapping
+    definition = function(x) x@strong_mapping
 )
 
 #' @title Get the mapping history for a BLASE Mapping Results object.
@@ -295,4 +329,22 @@ setMethod(
     f = "bootstrap_iterations",
     signature = "MappingResult",
     definition = function(x) x@bootstrap_iterations
+)
+
+#' @title Get the mapping history for a BLASE Mapping Results object.
+#'
+#' @concept mapping-result-object
+#'
+#' @rdname mapping-result-metric-getter
+#' @param x a [MappingResult] object
+#' @returns a String, the metric used to calculate the result.
+#' @export
+#' @inherit MappingResult examples
+setGeneric("metric", function(x) standardGeneric("metric"))
+
+#' @rdname mapping-result-metric-getter
+setMethod(
+  f = "metric",
+  signature = "MappingResult",
+  definition = function(x) x@metric
 )