+++ tests/testthat/helper.R
@@ -26,7 +26,7 @@ generate_test_mapping_result <- function(
     best_bin_ = 1,
     best_correlation_ = 0.05,
     top_2_distance_ = 0.001,
-    confident_mapping_ = FALSE,
+    strong_mapping_ = FALSE,
     bootstrap_iterations_ = 200,
     history_ = NULL) {
     if (is.null(history_)) {
@@ -42,7 +42,7 @@ generate_test_mapping_result <- function(
         best_bin = best_bin_,
         best_correlation = best_correlation_,
         top_2_distance = top_2_distance_,
-        confident_mapping = confident_mapping_,
+        strong_mapping = strong_mapping_,
         bootstrap_iterations = bootstrap_iterations_,
         history = history_
     ))